refactor(SequentialPaneDisplay): Merge wizard code and rename to wizard.
